feat: Clean log data for previous clients
The code changes add functionality to clean log data for previous clients. This is achieved by introducing the `LogRecoverer` and `LogWriter` classes. The `DataCleaner` class now manages clients and sends a clean flag to indicate the end of a client's data. The `LogRecoverer` class recovers client information from the log file, while the `LogWriter` class logs new and ended clients.

Note: The recent repository commits indicate a refactoring of the `QueryInfo` class to add a clean flag.