refactor: Add logging of previous clients in DataCleaner class
The code changes add logging of the previous clients in the `DataCleaner` class. This information is logged using the `logger.info` method and includes the list of previous clients retrieved from the `LogRecoverer` class. This logging is useful for tracking the clients that have already been processed and ignored during the log recovery and data cleaning process.

Note: The recent repository commits indicate a refactoring of the log recovery and data cleaning process to handle ignored client IDs.