refactor: Add sync functionality
This includes the addition of the `set_sync` and `get_sync` methods in the `DataFragment` class to manage the synchronization process. The `send_last` method is responsible for synchronizing with other nodes. This enhancement improves the synchronization and data transmission capabilities of the nodes.

Note: The recent repository commits indicate a refactoring of the synchronization process in the different nodes.